Add X connector contract entrypoint
